server/ai/carbon_trading_rag/requirements.txt
openai>=1.30.0
pinecone>=6.0.0
python-dotenv>=1.0.1
requests>=2.32.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
playwright>=1.51.0
PyMuPDF>=1.25.0
pandas>=2.2.2
tqdm>=4.66.4
rank-bm25>=0.2.2